requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
et-xmlfile==1.1.0
fake-useragent==0.1.11
free-proxy==1.0.2
idna==2.10
lxml==4.6.3
openpyxl==3.0.7
Pillow==8.2.0
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.4
